+++ example/next-app-router/src/app/layout.tsx
@@ -17,9 +17,7 @@ const Layout: FC<Props> = ({ children }) => {
       </head>
       <body>
         <KumaRegistry>
-          <k.div bg={"red"} className={inter.className}>
-            {children}
-          </k.div>
+          <div className={inter.className}>{children}</div>
         </KumaRegistry>
       </body>
     </html>