+++ example/vite/src/App.tsx
@@ -29,6 +29,9 @@ function App() {
   return (
     <Box color="GrayText" border="1px solid" borderColor="orange">
       test
+      <Styled>Styled</Styled>
+      <Text>日本語</Text>
+      <Dynamic />
     </Box>
   );
 }