+++ example/vite/vite.config.ts
@@ -2,13 +2,15 @@ import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react";
 import kumaUI from "@kuma-ui/vite";
 import Inspect from "vite-plugin-inspect";
-import { visualizer } from 'rollup-plugin-visualizer';
+import { visualizer } from "rollup-plugin-visualizer";
 
 // https://vitejs.dev/config/
 export default defineConfig({
   plugins: [
     react(),
-    kumaUI(),
+    kumaUI({
+      wasm: true,
+    }),
     Inspect({
       build: true,
       outputDir: ".inspect",
@@ -16,9 +18,7 @@ export default defineConfig({
   ],
   build: {
     rollupOptions: {
-      plugins: [
-        visualizer(),
-      ],
+      plugins: [visualizer()],
     },
   },
 });