+++ packages/babel-plugin/src/index.ts
@@ -1,7 +1,6 @@
 import type { PluginObj } from "@babel/core";
 import { visitor } from "./visitor";
 import type { Core } from "./core";
-import { transform } from "./transform";
 
 const plugin = (core: Core): PluginObj => {
   return {
@@ -24,5 +23,4 @@ const plugin = (core: Core): PluginObj => {
     visitor: visitor(core),
   };
 };
-export { transform };
 export default plugin;