+++ packages/core/src/css.ts
@@ -4,3 +4,6 @@ import { ThemeSystem } from "./theme";
 export const css = (_strings: TemplateStringsArray): string => {
   throw Error('Using the "css" in runtime is not supported.');
 };
+
+export const cx = (...classNames: (string | false | null | undefined)[]) =>
+  classNames.filter((className) => Boolean(className)).join(" ");