+++ packages/system/src/consumeStyleProps.test.ts
@@ -170,6 +170,14 @@ describe("consumeStyleProps function", () => {
       verticalAlign: "sub",
       userSelect: "none",
     });
+    consumeAndTest({
+      w: 8,
+      minW: "12px",
+      maxW: "2em",
+      h: "2em",
+      minH: "12px",
+      maxH: "2rem",
+    });
   });
   test("should consume list props", () => {
     consumeAndTest({