+++ packages/webpack-plugin/src/loader.ts
@@ -1,17 +1,25 @@
-import { transform } from "@kuma-ui/babel-plugin";
 import path from "path";
 import fs from "fs";
 import type { LoaderContext, RawLoaderDefinitionFunction } from "webpack";
 import { createHash } from "crypto";
 import { theme } from "@kuma-ui/sheet";
 import { getUserTheme } from "./getUserTheme";
-import { CSS_PATH } from "./plugin";
+import KumaUIWebpackPlugin, { CSS_PATH } from "./plugin";
+import { createRequire } from "module";
+import { compileSync } from "@kuma-ui/compiler";
 
 export const CSS_PARAM_NAME = "css";
 
+// tsup will replace __ESM__ with true during ESM build and false during CJS build when bundling.
+declare const __ESM__: boolean;
+const _require = __ESM__ ? createRequire(import.meta.url) : require;
+const emptyCssExtractionFile = _require.resolve("../assets/kuma.css");
+
 type Options = {
   config?: string;
-  cssOutputDir: string;
+  plugin: KumaUIWebpackPlugin;
+  outputDir?: string;
+  wasm?: boolean;
 };
 
 const kumaUiLoader: RawLoaderDefinitionFunction<Options> = function (
@@ -20,13 +28,13 @@ const kumaUiLoader: RawLoaderDefinitionFunction<Options> = function (
   // tell Webpack this loader is async
   const callback = this.async();
   const id = this.resourcePath;
-  const { config } = this.getOptions();
+  const { plugin, outputDir, wasm } = this.getOptions();
 
-  if (config) {
+  if (plugin.config) {
     // enable automatic rebuild for static theme props
     // <Box color={"colors.red.100"} />
-    this.addDependency(config);
-    const userTheme = getUserTheme(config);
+    this.addDependency(plugin.config);
+    const userTheme = getUserTheme(plugin.config);
     if (userTheme) {
       theme.setUserTheme(userTheme);
     }
@@ -43,20 +51,45 @@ const kumaUiLoader: RawLoaderDefinitionFunction<Options> = function (
 
   const outputPath = this._compiler?.options.output.path;
   if (!outputPath) throw Error("output path is not correctly set");
-  const result = transform(source.toString(), id);
+  const result = compileSync({ code: source.toString(), id, wasm });
   if (!result || !result.code) {
     callback(null, source);
     return;
   }
 
-  const css = (result.metadata as unknown as { css: string }).css || "";
+  const css = result.css || "";
 
   if (css) {
+    /**
+     * This is a temporary workaround to enable HMR (Hot Module Replacement) in Next.js client components.
+     * Currently, Client Component doesn't account for virtual files, so we need to emit an actual CSS file.
+     * TODO: Address and fix this issue.
+     */
+    if (outputDir && process.env.NODE_ENV !== "production") {
+      const codePrefix = fileLoader(css, {
+        context: this,
+        outputDir: outputDir,
+      });
+      callback(null, `${result.code}\n${codePrefix};`);
+      return;
+    }
+
     const params = new URLSearchParams({ [CSS_PARAM_NAME]: css });
 
     const importCSS = `import ${JSON.stringify(
-      `${this.utils.contextify(this.context, CSS_PATH)}?${params.toString()}`,
+      `${this.utils.contextify(
+        this.context,
+        emptyCssExtractionFile,
+      )}?${params.toString()}`,
     )};`;
+
+    if (plugin.watchMode) {
+      /**
+       * n Next.js version 13.5 and later, changes made in virtual files are no longer recognized by Next.js. Therefore, we need to emit random changes in the entry CSS file to ensure they are taken into account.
+       * @see {@link|https://github.com/vercel/next.js/discussions/59212}
+       */
+      fs.writeFileSync(CSS_PATH, `/* ${Date.now()} */`);
+    }
 
     callback(null, `${result.code}\n${importCSS};`);
     return;
@@ -65,3 +98,23 @@ const kumaUiLoader: RawLoaderDefinitionFunction<Options> = function (
 };
 
 export default kumaUiLoader;
+
+export function fileLoader(
+  src: string,
+  options: {
+    context: LoaderContext<unknown>;
+    outputDir: string;
+  },
+) {
+  const outDir = options.outputDir;
+  if (!fs.existsSync(outDir)) fs.mkdirSync(outDir);
+  const hash = createHash("md5").update(src).digest("hex");
+  const srcPath = path.posix.join(outDir, `${hash}.css`);
+  fs.writeFileSync(srcPath, src);
+
+  // Compute the relative path from the current file location to the output directory
+  const currentDir = options.context.context;
+  const relativeSrcPath = path.relative(currentDir, srcPath);
+
+  return `import "${relativeSrcPath}";`;
+}